+++ src/Config.hpp
@@ -13,25 +13,25 @@ using json = nlohmann::json;
 
 namespace Conf {
     /***** General parameters *****/
-    inline constexpr int WIDTH = 346; // px
-    inline constexpr int HEIGHT = 260; // px
+    inline constexpr size_t WIDTH = 346; // px
+    inline constexpr size_t HEIGHT = 260; // px
 
-    inline constexpr int EVENT_FREQUENCY = 1000; // μs
-    inline constexpr int DISPLAY_FREQUENCY = 30000; // μs
-    inline constexpr int UPDATE_PARAMETER_FREQUENCY = 1000000; // μs
+    inline constexpr size_t EVENT_FREQUENCY = 1000; // μs
+    inline constexpr size_t DISPLAY_FREQUENCY = 30000; // μs
+    inline constexpr size_t UPDATE_PARAMETER_FREQUENCY = 1000000; // μs
 
-    inline constexpr int TIME_WINDOW_SR = 20; // s
+    inline constexpr size_t TIME_WINDOW_SR = 20; // s
 
     inline const std::string CONF_FILE("/home/thomas/neuvisys-dv/configuration/conf.json");
     inline const std::string GUI_FILE("/home/thomas/neuvisys-dv/configuration/gui.json");
 }
 
 namespace Selection {
-    inline int LAYER = 0;
-    [[maybe_unused]] inline int LAYER2 = 0;
-    inline int SYNAPSE = 0;
-    inline unsigned int INDEX = 0;
-    inline unsigned int INDEX2 = 0;
+    inline size_t LAYER = 0;
+    [[maybe_unused]] inline size_t LAYER2 = 0;
+    inline size_t SYNAPSE = 0;
+    inline size_t INDEX = 0;
+    inline size_t INDEX2 = 0;
 }
 
 class NetworkConfig {
@@ -48,29 +48,29 @@ public:
     /***** Spiking Neural Network layout parameters *****/
     std::string Neuron1Config;
     std::string Neuron2Config;
-    int L1Width{};
-    int L1Height{};
-    int L1Depth{};
-    int L2Width{};
-    int L2Height{};
-    int L2Depth{};
+    size_t L1Width{};
+    size_t L1Height{};
+    size_t L1Depth{};
+    size_t L2Width{};
+    size_t L2Height{};
+    size_t L2Depth{};
 
-    std::vector<int> L1XAnchor;
-    std::vector<int> L1YAnchor;
-    std::vector<int> L2XAnchor;
-    std::vector<int> L2YAnchor;
-    int Neuron1Width{};
-    int Neuron1Height{};
-    int Neuron1Synapses{};
+    std::vector<size_t> L1XAnchor;
+    std::vector<size_t> L1YAnchor;
+    std::vector<size_t> L2XAnchor;
+    std::vector<size_t> L2YAnchor;
+    size_t Neuron1Width{};
+    size_t Neuron1Height{};
+    size_t Neuron1Synapses{};
 
-    int Neuron2Width{};
-    int Neuron2Height{};
+    size_t Neuron2Width{};
+    size_t Neuron2Height{};
     bool WeightSharing{};
 };
 
 class NeuronConfig {
 public:
-    NeuronConfig(std::string configFile, int type);
+    NeuronConfig(std::string configFile, size_t type);
 /***** Neurons internal parameters *****/
     double TAU_M{}; // μs
     double TAU_LTP{}; // μs
@@ -88,7 +88,7 @@ public:
     double VRESET{}; // mV
     double VTHRESH{}; // mV
 
-    long SYNAPSE_DELAY{}; // μs
+    size_t SYNAPSE_DELAY{}; // μs
 
     double NORM_FACTOR{};
     double DECAY_FACTOR{};
