+++ src/Event.hpp
@@ -5,33 +5,33 @@
 
 class Event {
     long m_timestamp;
-    int m_x;
-    int m_y;
+    size_t m_x;
+    size_t m_y;
     bool m_polarity;
-    int m_synapse;
+    size_t m_synapse;
 public:
     Event() = default;
-    inline Event(long timestamp, int x, int y, bool polarity) : m_timestamp(timestamp), m_x(x), m_y(y), m_polarity(polarity) {m_synapse = 0;}
-    inline Event(long timestamp, int x, int y, bool polarity, int synapse) : m_timestamp(timestamp), m_x(x), m_y(y), m_polarity(polarity), m_synapse(synapse) {}
+    inline Event(long timestamp, size_t x, size_t y, bool polarity) : m_timestamp(timestamp), m_x(x), m_y(y), m_polarity(polarity) {m_synapse = 0;}
+    inline Event(long timestamp, size_t x, size_t y, bool polarity, size_t synapse) : m_timestamp(timestamp), m_x(x), m_y(y), m_polarity(polarity), m_synapse(synapse) {}
     [[nodiscard]] inline long timestamp() const {return m_timestamp;}
-    [[nodiscard]] inline int x() const {return m_x;}
-    [[nodiscard]] inline int y() const {return m_y;}
+    [[nodiscard]] inline size_t x() const {return m_x;}
+    [[nodiscard]] inline size_t y() const {return m_y;}
     [[nodiscard]] inline bool polarity() const {return m_polarity;}
-    [[nodiscard]] inline int synapse() const {return m_synapse;}
+    [[nodiscard]] inline size_t synapse() const {return m_synapse;}
 };
 
 class NeuronEvent {
     long m_timestamp;
-    int m_x;
-    int m_y;
-    int m_z;
+    size_t m_x;
+    size_t m_y;
+    size_t m_z;
 public:
     NeuronEvent() = default;
-    inline NeuronEvent(long timestamp, int x, int y, int z) : m_timestamp(timestamp), m_x(x), m_y(y), m_z(z) {}
+    inline NeuronEvent(long timestamp, size_t x, size_t y, size_t z) : m_timestamp(timestamp), m_x(x), m_y(y), m_z(z) {}
     [[nodiscard]] inline long timestamp() const {return m_timestamp;}
-    [[nodiscard]] inline int x() const {return m_x;}
-    [[nodiscard]] inline int y() const {return m_y;}
-    [[nodiscard]] inline int z() const {return m_z;}
+    [[nodiscard]] inline size_t x() const {return m_x;}
+    [[nodiscard]] inline size_t y() const {return m_y;}
+    [[nodiscard]] inline size_t z() const {return m_z;}
 };
 
 #endif //NEUVISYS_DV_EVENT_HPP
