+++ src/neurons/SimpleNeuron.hpp
@@ -19,12 +19,12 @@ class SimpleNeuron : public Neuron {
     boost::circular_buffer<Event> m_events;
     std::priority_queue<Event, std::vector<Event>, CompareEventsTimestamp> m_waitingList;
 public:
-    SimpleNeuron(size_t index, NeuronConfig &conf, Luts &luts, Position pos, Position offset, xt::xarray<double> &weights, int nbSynapses);
-    void newEvent(long timestamp, int x, int y, bool polarity) override;
+    SimpleNeuron(size_t index, NeuronConfig &conf, Luts &luts, Position pos, Position offset, xt::xarray<double> &weights, size_t nbSynapses);
+    void newEvent(long timestamp, size_t x, size_t y, bool polarity) override;
     void update(long time) override;
-    double getWeights(int p, int s, int x, int y) override;
+    double getWeights(size_t p, size_t s, size_t x, size_t y) override;
 private:
-    void membraneUpdate(long timestamp, int x, int y, bool polarity, int synapse);
+    void membraneUpdate(long timestamp, size_t x, size_t y, bool polarity, size_t synapse);
     void spike(long time);
     void updateSTDP();
     void normalizeWeights();
